hw3 grading

+++ src/homework/03_iteration/dna.h
@@ -7,7 +7,7 @@ using std::cout; using std::cin; using std::string;
 Write prototype for function get_gc_content that accepts
 a const reference string parameter and returns a double
 */
-int get_gc_content(const string &dna);
+double get_gc_content(const string &dna);
 
 
 /*
